Change: Split init functions
