+++ src/codegen/mod.rs
@@ -1,4 +1,5 @@
 pub mod compilation_result;
 pub mod compile;
 pub mod error;
+pub mod initialization;
 pub mod target_machine;
